pyproject.toml
[tool.poetry]
name = "nexus"
version = "0.1.0"
description = "NEXUS Platform"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13" # Assuming Python 3.11 based on CI/CD
fastapi = "==0.104.1"
uvicorn = {extras = ["standard"], version = "==0.24.0"}
pydantic = {extras = ["email"], version = "==2.5.0"}
pydantic-settings = "==2.1.0"
sqlalchemy = "==2.0.23"
asyncpg = "==0.29.0"
alembic = "==1.13.1"
psycopg2-binary = "==2.9.9"
redis = "==5.0.1"
python-jose = {extras = ["cryptography"], version = "==3.3.0"}
passlib = {extras = ["bcrypt"], version = "==1.7.4"}
python-multipart = "==0.0.6"
cryptography = "==42.0.8"
bleach = "==6.0.0"
httpx = "==0.25.2"
websockets = "==12.0"
aiohttp = "==3.9.1"
pandas = "==2.1.4"
numpy = "==1.25.2"
openpyxl = "==3.1.2"
openai = "==1.3.7"
anthropic = "==0.7.8"
scikit-learn = "==1.3.2"
prometheus-client = "==0.19.0"
structlog = "==23.2.0"
sentry-sdk = {extras = ["fastapi"], version = "==1.38.0"}
python-dotenv = "==1.0.0"
celery = "==5.3.4"
flower = "==2.0.1"
gunicorn = "==21.2.0"
requests = "==2.31.0"
pyyaml = "==6.0.1"
toml = "==0.10.2"
aioredis = "==2.0.1"
kombu = "==5.3.4"
click = "==8.1.7"
rich = "==13.7.0"
typer = "==0.9.0"
tenacity = "==8.2.3"
fastapi-swagger-ui = "==0.1.0"
fastapi-redoc = "==0.1.0"
python-magic = "==0.4.27"
pillow = "==10.1.0"
fastapi-mail = "==1.4.1"
python-dateutil = "==2.8.2"
pytz = "==2023.3"
orjson = "==3.9.10"
msgpack = "==1.0.7"
gzip-stream = "==0.1.0"
psutil = "==5.9.6"
python-decouple = "==3.8"

[tool.poetry.group.dev.dependencies]
pytest = "==7.4.3"
pytest-asyncio = "==0.21.1"
pytest-cov = "==4.1.0"
black = "==23.11.0"
isort = "==5.12.0"
flake8 = "==6.1.0"
mypy = "==1.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
